examples/requirements.txt
--index-url https://test.pypi.org/simple/
--extra-index-url https://pypi.org/simple
amigo-python-sdk
python-dotenv>=1.0,<2.0
